feat(gtk): Implemented the GTK Firmware Widget
Implements the firmware widget library, which is written in Rust; with
optional C FFI support, to enable integration in GNOME Settings.
Examples for the Rust and C APIs are available.

Currently, only Thelio and Thelio I/O firmware updates are supported. In
the future, integration with the LVFS firmware service is possible.