feat(core): Use slotmap EC arch to manage device widgets
System, device, and Thelio I/O widgets are now stored in a single
storage, referenced by an entity ID. This allows the error-handling
interface to be aware of which entity triggered a failed update,
so that the progress bar can be switched back to a button.

Additionally, this enables for Thelio I/O widgets to be treated
uniquely, where only the first widget will have a clickable button,
if any of the devices are updateable, and where a succcessful update
shall update the labels of all widgets.