feat(notify): Create firmware-manager-notify package
There is now a `firmware-manager-notify` debian package, which contains
a `com.system76.FirmwareManager.Notify` binary, and an autostart desktop
file to check for firmware updates.

When updates are found, a desktop notification will be displayed. If
GNOME Control Center integration is discovered, the firmware panel in
GNOME Control Center will be displayed. If not, the standalone firmware
manager application will be called.