fix(core): Add generic version collection sort to fix fwupd release sorting
We were relying on `BTreeSet` to sort releases from fwupd, but this only
accounts for alphabetical sorting, rather than version sorting. To fix
this, the core library now contains a private `Versioned` trait, along
with a generic sorting function that operates on a mutable slice of
`Versioned` types.

This trait is implemented on `fwupd-bus`'s `Release` type, so that we can
now guarantee that releases are sorted correctly. A unit test was also
added to ensure that this solution is working correctly. Additionally, this
required updating the `fwupd-dbus` crate to switch back to storing releases
in a `Vec`.

Fixes #32