feat(core): Add usb_hotplug_event_loop convenience function
Adds a convenience function which a frontend can use for responding to
hotplug events. This should mainly be used to send the
`FirmwareEvent::Scan` signal to the background thread. A better solution
will be available once the async/await syntax is made official in Rust
1.39.0.

First half of the solution for #49